test(review): isolate evidence escaping coverage for independent landing
